restructured (again, reverting) the encoded_file_reader. also added a wrapper file_handler.rs to take care of what the restructure was initially for
reason: I realize I have to keep track of the merkle tree if I want this to be efficient and not re-calculate it every time the file is dropped/instantiated. Saving the merkle tree at the end doesn't realy work since I need to use the length of the file to calculate the dimensions of the matrix. The only real work around is to have a parent function that keeps track of all the state.